+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "css-minifier-tool-and-compressor" extension will be 
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
+## [0.0.2] - 2024-06-08
+
+- Add explorer context menu to minify css & js files
+- Add prefix options to minify css & js files
+
+## [0.0.1] - 2024-06-08
 
 - Initial release